Complete Case Study File
A second Github repo of mine (CaseStudy_Akshay) contains all the details of the Case work submitted on the last day and the link for the same is inside this file. Along with that the file contains detailed screenshots of all the tasks which were asked to cover under this case work.